coop with errors

+++ src/main/java/SmartBuildingResources/Server/Resource/raw/CameraRawSensor.java
@@ -12,13 +12,13 @@ public class CameraRawSensor extends SmartObjectResource<Integer>{
 
     private static Logger logger = LoggerFactory.getLogger(CameraRawSensor.class);
 
-    private static final int MIN_VALUE = 0;
+    private static final double MIN_VALUE = 0;
 
-    private static final int MAX_VALUE = 30;
+    private static final double MAX_VALUE = 30;
 
-    private static final int MIN_VARIATION = 0;
+    private static final double MIN_VARIATION = 0;
 
-    private static final int MAX_VARIATION = 5;
+    private static final double MAX_VARIATION = 5;
 
     private static final String LOG_DISPLAY_NAME = "CameraSensor";
 
@@ -50,7 +50,7 @@ public class CameraRawSensor extends SmartObjectResource<Integer>{
         try{
 
             this.random = new Random(System.currentTimeMillis());
-            this.updatedValue = MIN_VALUE + this.random.nextInt()*(MAX_VALUE - MIN_VALUE);
+            this.updatedValue = (Integer) (MIN_VALUE + this.random.nextDouble()*(MAX_VALUE - MIN_VALUE));
 
             startPeriodicEventValueUpdateTask();
